# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "QueueLess API",
    "description": "Predict waiting time at government offices + remote virtual tokens",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/devices": {
      "post": {
        "tags": [
          "Core V1"
        ],
        "summary": "Register Device",
        "description": "C7: Register FCM device token and language for push notifications.\nUpserts into devices table.",
        "operationId": "register_device_v1_devices_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceRegisterIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/me": {
      "patch": {
        "tags": [
          "Core V1"
        ],
        "summary": "Update Profile",
        "description": "C11: Update profile / language. Server notification templates follow it.",
        "operationId": "update_profile_v1_me_patch",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdateIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/display/{office_id}": {
      "get": {
        "tags": [
          "Core V1"
        ],
        "summary": "Get Public Lobby Display",
        "description": "Public lobby board: counter label and now-serving display code only.\nNo personal data exposed. No authentication required.",
        "operationId": "get_public_lobby_display_v1_display__office_id__get",
        "parameters": [
          {
            "name": "office_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Office Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisplayBoardOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/queues/{service_id}/pause": {
      "post": {
        "tags": [
          "Core V1"
        ],
        "summary": "Pause Queue Booking",
        "description": "O10: Pause booking for a service. Reason required. Sets queue_state.paused = True.",
        "operationId": "pause_queue_booking_v1_queues__service_id__pause_post",
        "parameters": [
          {
            "name": "service_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Service Id"
            }
          },
          {
            "name": "reason",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 2,
              "description": "Reason for pausing booking",
              "title": "Reason"
            },
            "description": "Reason for pausing booking"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Core V1"
        ],
        "summary": "Resume Queue Booking",
        "description": "O10: Resume booking for a service. Sets queue_state.paused = False.",
        "operationId": "resume_queue_booking_v1_queues__service_id__pause_delete",
        "parameters": [
          {
            "name": "service_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Service Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeviceRegisterIn": {
        "properties": {
          "fcm_token": {
            "type": "string",
            "title": "Fcm Token"
          },
          "platform": {
            "type": "string",
            "title": "Platform",
            "default": "android"
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "en"
          }
        },
        "type": "object",
        "required": [
          "fcm_token"
        ],
        "title": "DeviceRegisterIn"
      },
      "DisplayBoardOut": {
        "properties": {
          "office_id": {
            "type": "string",
            "title": "Office Id"
          },
          "office_name": {
            "type": "string",
            "title": "Office Name"
          },
          "counters": {
            "items": {
              "$ref": "#/components/schemas/DisplayCounterOut"
            },
            "type": "array",
            "title": "Counters"
          }
        },
        "type": "object",
        "required": [
          "office_id",
          "office_name",
          "counters"
        ],
        "title": "DisplayBoardOut"
      },
      "DisplayCounterOut": {
        "properties": {
          "counter_label": {
            "type": "string",
            "title": "Counter Label"
          },
          "now_serving": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Now Serving"
          }
        },
        "type": "object",
        "required": [
          "counter_label"
        ],
        "title": "DisplayCounterOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProfileUpdateIn": {
        "properties": {
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "type": "object",
        "title": "ProfileUpdateIn"
      },
...
      "SuccessResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Operation completed successfully"
          }
        },
        "type": "object",
        "title": "SuccessResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}